Services seperation api and web
