+++ internal/services/message_service.go
@@ -1,0 +1,1 @@
+package services
