+++ internal/web/handlers.go
@@ -1,73 +1,57 @@
-package main
+package web
 
 import (
 	"errors"
-	"html/template"
 	"net/http"
+	"time"
+
+	"devops/minitwit/api_models"
+	"devops/minitwit/internal/auth"
+	"devops/minitwit/internal/models"
+	"devops/minitwit/internal/services"
+	"devops/minitwit/internal/session"
 
 	"github.com/gorilla/mux"
 	"github.com/rs/zerolog/log"
 	"gorm.io/gorm"
 )
 
-var baseTpl = template.Must(
-	template.New("base").Funcs(funcMap).ParseFiles("templates/layout.html"),
-)
+const PER_PAGE = 30
 
-var loginTpl = template.Must(
-	template.Must(baseTpl.Clone()).ParseFiles("templates/login.html"),
-)
+func convertAPIMessagesToTemplateMessages(apiMsgs []api_models.Message) []map[string]any {
+	msgs := make([]map[string]any, 0, len(apiMsgs))
 
-var registerTpl = template.Must(
-	template.Must(baseTpl.Clone()).ParseFiles("templates/register.html"),
-)
-var timelineTpl = template.Must(
-	template.Must(baseTpl.Clone()).Funcs(funcMap).ParseFiles("templates/timeline.html"),
-)
-
-var funcMap = template.FuncMap{
-	"url": func(urlName string) string {
-		return routes[urlName]
-	},
-	"gravatar": gravatar_url,
-	"datetime": func(ts any) string {
-		switch v := ts.(type) {
-		case int64:
-			return FormatDatetime(v)
-		case int:
-			return FormatDatetime(int64(v))
-		default:
-			return ""
-		}
-	},
-}
-
-var routes = map[string]string{
-	"timeline":        "/",
-	"login":           "/login",
-	"public_timeline": "/public",
-	"register":        "/register",
-	"logout":          "/logout",
-	// TODO: extend with all name -> api route
-}
-
-// TODO right now the password is matched agains exactly what is in the db, should be hash
-func Login(w http.ResponseWriter, r *http.Request) {
-
-	// Redirect the user if they are already logged in.
-	_, ok := TryGetUserFromRequest(r)
-	if ok {
-		http.Redirect(w, r, "/public", http.StatusFound)
+	for _, m := range apiMsgs {
+		msgs = append(msgs, map[string]any{
+			"text":     m.Content,
+			"username": m.User,
+			"pub_date": 0,
+		})
 	}
 
-	data := Data{
+	return msgs
+}
+
+var database *gorm.DB
+
+func SetDB(db *gorm.DB) {
+	database = db
+}
+
+func Login(w http.ResponseWriter, r *http.Request) {
+	_, ok := auth.TryGetUserFromRequest(r)
+	if ok {
+		http.Redirect(w, r, "/public", http.StatusFound)
+		return
+	}
+
+	data := models.Data{
 		Error:        "",
 		FormUsername: "",
-		Flashes:      GetFlashes(w, r),
+		Flashes:      session.GetFlashes(w, r),
 		User:         nil,
 	}
 
-	// On GET request we return the template.
 	if r.Method == http.MethodGet {
 		if err := loginTpl.ExecuteTemplate(w, "layout", data); err != nil {
 			http.Error(w, err.Error(), http.StatusInternalServerError)
@@ -75,59 +59,53 @@ func Login(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	// Get username and password from the template form.
 	username := r.FormValue("username")
 	password := r.FormValue("password")
 
-	userUser, errorMessage := ValidateLogin(username, password)
+	userUser, errorMessage := services.ValidateLogin(username, password)
 
-	// Add user to session and redirect
 	if userUser != nil {
-
-		session, err := store.Get(r, "session")
+		sessionData, err := session.GetStore().Get(r, "session")
 		if err != nil {
 			http.Error(w, "Session error", http.StatusInternalServerError)
 			return
 		}
 
-		session.Values["user_id"] = userUser.User_id
+		sessionData.Values["user_id"] = userUser.User_id
 
-		err = session.Save(r, w)
+		err = sessionData.Save(r, w)
 		if err != nil {
 			http.Error(w, "Could not save session", http.StatusInternalServerError)
 			return
 		}
-		AddFlash(w, r, "You were logged in")
+		session.AddFlash(w, r, "You were logged in")
 		http.Redirect(w, r, "/public", http.StatusSeeOther)
 		return
 	}
-	// If the username or password is wrong display error in login page.
+
 	data.Error = errorMessage
 	if err := loginTpl.ExecuteTemplate(w, "layout", data); err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 	}
 }
 
-// TODO: FollowUser(username)
 func FollowUser(w http.ResponseWriter, r *http.Request) {
-	user, ok := TryGetUserFromRequest(r)
-
+	user, ok := auth.TryGetUserFromRequest(r)
 	if !ok {
 		http.Error(w, "Unauthorized", http.StatusUnauthorized)
-		//http.Redirect(w, r, "/public", http.StatusFound)
 		return
 	}
 
 	vars := mux.Vars(r)
 	usernameToFollow := vars["username"]
 
-	whomID := get_user_id(usernameToFollow)
+	whomID := services.GetUserID(usernameToFollow)
 	if whomID == -1 {
 		http.Error(w, "User not found", http.StatusNotFound)
 		return
 	}
 
-	follower := Follower{
+	follower := models.Follower{
 		Who_id:  user.User_id,
 		Whom_id: whomID,
 	}
@@ -137,59 +115,58 @@ func FollowUser(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, "Failed to follow user: "+res.Error.Error(), http.StatusInternalServerError)
 		return
 	}
-	// Add the flash message to the session:
-	AddFlash(w, r, "You are now following \""+usernameToFollow+"\"")
+
+	session.AddFlash(w, r, "You are now following \""+usernameToFollow+"\"")
 	http.Redirect(w, r, "/"+usernameToFollow, http.StatusSeeOther)
 }
 
-// TODO: UnfollowUser(username)
 func UnfollowUser(w http.ResponseWriter, r *http.Request) {
-	user, ok := TryGetUserFromRequest(r)
+	user, ok := auth.TryGetUserFromRequest(r)
 	if !ok {
 		http.Error(w, "Unauthorized", http.StatusUnauthorized)
-		//http.Redirect(w, r, "/public", http.StatusFound)
 		return
 	}
+
 	vars := mux.Vars(r)
 	usernameToUnfollow := vars["username"]
 
-	whomID := get_user_id(usernameToUnfollow)
+	whomID := services.GetUserID(usernameToUnfollow)
 	if whomID == -1 {
 		http.Error(w, "User not found", http.StatusNotFound)
 		return
 	}
 
-	res := database.Where("who_id = ? AND whom_id = ?", user.User_id, whomID).Delete(&Follower{})
+	res := database.Where("who_id = ? AND whom_id = ?", user.User_id, whomID).Delete(&models.Follower{})
 	if res.Error != nil {
 		http.Error(w, "Failed to unfollow user: "+res.Error.Error(), http.StatusInternalServerError)
 		return
 	}
-	AddFlash(w, r, "You are no longer following \""+usernameToUnfollow+"\"")
+
+	session.AddFlash(w, r, "You are no longer following \""+usernameToUnfollow+"\"")
 	http.Redirect(w, r, "/"+usernameToUnfollow, http.StatusSeeOther)
 }
 
-// Removes user_id from session and redirects to "/"
 func Logout(w http.ResponseWriter, r *http.Request) {
-
-	session, err := store.Get(r, "session")
+	sessionData, err := session.GetStore().Get(r, "session")
 	if err != nil {
 		http.Error(w, "session error", http.StatusInternalServerError)
 		return
 	}
 
-	delete(session.Values, "user_id")
+	delete(sessionData.Values, "user_id")
 
-	err = session.Save(r, w)
+	err = sessionData.Save(r, w)
 	if err != nil {
 		http.Error(w, "could not save session", http.StatusInternalServerError)
 		return
 	}
-	AddFlash(w, r, "You were logged out")
+
+	session.AddFlash(w, r, "You were logged out")
 	http.Redirect(w, r, "/", http.StatusSeeOther)
 }
 
 func MyTimeline(w http.ResponseWriter, r *http.Request) {
-	user, ok := TryGetUserFromRequest(r)
+	user, ok := auth.TryGetUserFromRequest(r)
 	if !ok {
 		http.Redirect(w, r, "/public", http.StatusFound)
 		return
@@ -203,7 +180,7 @@ func MyTimeline(w http.ResponseWriter, r *http.Request) {
 		Joins(`JOIN "user" u ON m.author_id = u.user_id`).
 		Where("m.flagged = 0 AND (u.user_id = ? OR u.user_id IN (?) )",
 			user.User_id,
-			database.Model(&Follower{}).Select("whom_id").Where("who_id = ?", user.User_id),
+			database.Model(&models.Follower{}).Select("whom_id").Where("who_id = ?", user.User_id),
 		).
 		Order("m.pub_date DESC").
 		Limit(PER_PAGE).
@@ -214,11 +191,11 @@ func MyTimeline(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	data := Data{
+	data := models.Data{
 		Messages: msgs,
-		Endpoint: "timeline", // Add this line
+		Endpoint: "timeline",
 		User:     &user,
-		Flashes:  GetFlashes(w, r),
+		Flashes:  session.GetFlashes(w, r),
 	}
 
 	if err := timelineTpl.ExecuteTemplate(w, "layout", data); err != nil {
@@ -226,14 +203,13 @@ func MyTimeline(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
-
 }
 
 func UserTimeline(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	username := vars["username"]
 
-	flashes := GetFlashes(w, r)
+	flashes := session.GetFlashes(w, r)
 
 	var messages []map[string]any
 
@@ -251,9 +227,8 @@ func UserTimeline(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var profileUser User
+	var profileUser models.User
 
-	// You need to get the profile user data
 	res = database.
 		Select("user_id, username").
 		Where("username = ?", username).
@@ -268,36 +243,28 @@ func UserTimeline(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	data := Data{
+	data := models.Data{
 		FormUsername: username,
-		ProfileUser:  &profileUser, // Add this
+		ProfileUser:  &profileUser,
 		Messages:     messages,
-		Endpoint:     "user_timeline", // Add this
-		// You also need to set Followed based on whether the current user follows this user
-		Followed: false, // Set this appropriately
-		Flashes:  flashes,
+		Endpoint:     "user_timeline",
+		Followed:     false,
+		Flashes:      flashes,
 	}
 
-	user, ok := TryGetUserFromRequest(r)
-
+	user, ok := auth.TryGetUserFromRequest(r)
 	if ok {
 		data.User = &user
 	}
 
-	var follower Follower
+	var follower models.Follower
 
 	res = database.
 		Select("whom_id").
-		Where("who_id = ? AND whom_id = ?", user.User_id, get_user_id(profileUser.Username)).
+		Where("who_id = ? AND whom_id = ?", user.User_id, services.GetUserID(profileUser.Username)).
 		First(&follower)
 
-	if res.Error != nil {
-		if errors.Is(res.Error, gorm.ErrRecordNotFound) {
-			// Not following
-		} else {
-			log.Warn().Stack().Err(res.Error).Msg("")
-		}
-	} else {
+	if res.Error == nil {
 		data.Followed = true
 	}
 
@@ -308,29 +275,23 @@ func UserTimeline(w http.ResponseWriter, r *http.Request) {
 }
 
 func Timeline(w http.ResponseWriter, r *http.Request) {
+	client := NewAPIClient()
 
-	var msgs []map[string]any
-
-	res := database.
-		Table("message").
-		Select(`message.message_id, message.text, message.pub_date, u.username`).
-		Joins(`JOIN "user" u ON u.user_id = message.author_id`).
-		Order("message.pub_date DESC").
-		Limit(PER_PAGE).
-		Find(&msgs)
-
-	if res.Error != nil {
-		http.Error(w, res.Error.Error(), http.StatusInternalServerError)
+	apiMsgs, err := client.GetPublicMessages()
+	if err != nil {
+		http.Error(w, "Could not load public timeline from API: "+err.Error(), http.StatusBadGateway)
 		return
 	}
 
-	data := Data{
-		Messages: msgs,
-		Endpoint: "public_timeline", // Add this line
-		Flashes:  GetFlashes(w, r),
-	}
-	user, ok := TryGetUserFromRequest(r)
+	msgs := convertAPIMessagesToTemplateMessages(apiMsgs)
 
+	data := models.Data{
+		Messages: msgs,
+		Endpoint: "public_timeline",
+		Flashes:  session.GetFlashes(w, r),
+	}
+
+	user, ok := auth.TryGetUserFromRequest(r)
 	if ok {
 		data.User = &user
 	}
@@ -339,17 +300,16 @@ func Timeline(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
-
 }
 
 func Register(w http.ResponseWriter, r *http.Request) {
-
 	if r.Method == http.MethodGet {
 		if err := registerTpl.ExecuteTemplate(w, "layout", nil); err != nil {
 			http.Error(w, err.Error(), http.StatusInternalServerError)
 		}
 		return
 	}
+
 	if err := r.ParseForm(); err != nil {
 		http.Error(w, "Invalid form", http.StatusBadRequest)
 		return
@@ -360,9 +320,9 @@ func Register(w http.ResponseWriter, r *http.Request) {
 	firstPassword := r.FormValue("password")
 	secondPassword := r.FormValue("password2")
 
-	data := Data{}
+	data := models.Data{}
 
-	ok, registerError := ValidateRegister(username, email, firstPassword, secondPassword)
+	ok, registerError := services.ValidateRegister(username, email, firstPassword, secondPassword)
 
 	data.Error = registerError
 	if !ok {
@@ -370,9 +330,9 @@ func Register(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var pwHash, err = HashPassword(firstPassword)
+	pwHash, err := services.HashPassword(firstPassword)
 
-	user := User{
+	user := models.User{
 		Username: username,
 		Email:    email,
 		Pw_hash:  pwHash,
@@ -380,28 +340,32 @@ func Register(w http.ResponseWriter, r *http.Request) {
 
 	res := database.Create(&user)
 	if res.Error != nil {
-		data = Data{Error: "Failed to register: " + err.Error(), FormUsername: username}
+		data = models.Data{Error: "Failed to register: " + err.Error(), FormUsername: username}
 		registerTpl.ExecuteTemplate(w, "layout", data)
 		return
 	}
-	AddFlash(w, r, "You were successfully registered and can login now")
+
+	session.AddFlash(w, r, "You were successfully registered and can login now")
 	http.Redirect(w, r, "/login", http.StatusSeeOther)
 }
 
-func RegisterRoutes(router *mux.Router) {
-	router.HandleFunc("/", MyTimeline).Methods("GET")
+func AddMessage(w http.ResponseWriter, r *http.Request) {
+	user, ok := auth.TryGetUserFromRequest(r)
+	if !ok {
+		http.Error(w, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
 
-	router.HandleFunc("/public", Timeline).Methods("GET")
+	res := database.Create(&models.Message{
+		Author_id: user.User_id,
+		Text:      r.FormValue("text"),
+		Pub_date:  int(time.Now().Unix()),
+	})
+	if res.Error != nil {
+		http.Error(w, "Failed post message: "+res.Error.Error(), http.StatusInternalServerError)
+		return
+	}
 
-	router.HandleFunc("/add_message", AddMessage).Methods("POST")
-
-	router.HandleFunc("/login", Login).Methods("GET", "POST")
-
-	router.HandleFunc("/logout", Logout)
-
-	router.HandleFunc("/register", Register).Methods("GET", "POST")
-
-	router.HandleFunc("/{username}/follow", FollowUser).Methods("GET")
-	router.HandleFunc("/{username}/unfollow", UnfollowUser).Methods("GET")
-	router.HandleFunc("/{username}", UserTimeline).Methods("GET")
+	session.AddFlash(w, r, "Your message was recorded")
+	http.Redirect(w, r, "/", http.StatusSeeOther)
 }
