+++ internal/web/sessions.go
@@ -1,0 +1,1 @@
+package web
